Move trigonometry methods to their own class in MathUtils
