+++ GW2EIEvtcParser/ParsedData/CombatEvents/StatusEvents/EffectEvents/AbstractEffectEvent.cs
@@ -1,4 +1,5 @@
 ﻿using System.Numerics;
+using static GW2EIEvtcParser.EIData.Trigonometry;
 using static GW2EIEvtcParser.ParserHelper;
 
 namespace GW2EIEvtcParser.ParsedData;
